Fixed styles, remove code repetition
