+++ packages/app-extension/src/app/Router.tsx
@@ -42,6 +42,7 @@ import {
 } from "../components/Unlocked/Approvals/ApproveTransaction";
 import { WithAuth } from "../components/Unlocked/WithAuth";
 import { refreshFeatureGates } from "../gates/FEATURES";
+import { sanitizeTransactionWithFeeConfig } from "../utils/solana";
 
 import "./App.css";
 
@@ -211,27 +212,7 @@ function QueryApproveTransaction() {
               });
               return;
             }
-            let tx = deserializeTransaction(txStr);
-            if (feeConfig && feeConfig.priorityFee && tx.version === "legacy") {
-              const transaction = deserializeLegacyTransaction(txStr);
-              const modifyComputeUnits =
-                ComputeBudgetProgram.setComputeUnitLimit({
-                  units: feeConfig.computeUnits,
-                });
-
-              const addPriorityFee = ComputeBudgetProgram.setComputeUnitPrice({
-                microLamports: feeConfig.priorityFee,
-              });
-
-              transaction.add(modifyComputeUnits);
-              transaction.add(addPriorityFee);
-              tx = deserializeTransaction(
-                bs58.encode(
-                  transaction.serialize({ requireAllSignatures: false })
-                )
-              );
-            }
-
+            const tx = sanitizeTransactionWithFeeConfig(txStr, feeConfig);
             await background.response({
               id: requestId,
               result: {
