+++ packages/react-common/src/components/base/Inputs.tsx
@@ -280,7 +280,7 @@ export const SmallInput = ({
           textAlign: "right",
           padding: 2,
           background: theme.custom.colors.background,
-          borderRadius: 0,
+          borderRadius: 5,
           outline: "none",
           color: theme.custom.colors.fontColor2,
         },
@@ -289,7 +289,7 @@ export const SmallInput = ({
         marginBottom: 5,
         width: 100,
         padding: 0,
-        borderRadius: 0,
+        borderRadius: 5,
         border: "none",
         outline: "none",
       }}
